Behavior refactoring
