+++ src/test/java/br/unicamp/meca/mind/behavior/Test1AndTest2BehaviorCodelet.java
@@ -7,6 +7,7 @@ import java.util.ArrayList;
 
 import br.unicamp.cst.core.entities.Memory;
 import br.unicamp.meca.models.ActionSequencePlan;
+import br.unicamp.meca.models.ActionStep;
 import br.unicamp.meca.system1.codelets.BehaviorCodelet;
 
 /**
@@ -16,20 +17,18 @@ import br.unicamp.meca.system1.codelets.BehaviorCodelet;
 public class Test1AndTest2BehaviorCodelet extends BehaviorCodelet {
 
 	public Test1AndTest2BehaviorCodelet(String id, ArrayList<String> perceptualCodeletsIds,
-			ArrayList<String> motivationalCodeletsIds, String soarCodeletId, ActionSequencePlan actionSequencePlan) {
-		super(id, perceptualCodeletsIds, motivationalCodeletsIds, soarCodeletId, actionSequencePlan);
+			ArrayList<String> motivationalCodeletsIds, String soarCodeletId) {
+		super(id, perceptualCodeletsIds, motivationalCodeletsIds, soarCodeletId);
 	}
 
 	@Override
-	public void trackActionSequencePlan(ArrayList<Memory> perceptualMemories, ActionSequencePlan actionSequencePlan) {
+	protected ActionSequencePlan buildActionSequencePlan(ArrayList<Memory> perceptualMemories) {
 		
-		if(actionSequencePlan == null || actionSequencePlan.getActionStepSequence() == null) {
-			return;
-		}
-					
-		actionSequencePlan.setCurrentActionIdIndex(0);
-		
-		//In this test, will never go on to the second action. We could have limit on perceptual memory to move on, though.
+		ActionStep as1 = new ActionStep("Test1ActionFromPlanningCodelet");
+        ActionStep as2 = new ActionStep("Test2ActionFromPlanningCodelet");
+        ActionSequencePlan test1Test2ActionSequence = new ActionSequencePlan(new ActionStep[] {as1,as2});
+        
+		return test1Test2ActionSequence;
 	}
 
 }
